Fix warnings

+++ Source/Initialization/Nyx_initdata.cpp
@@ -266,7 +266,7 @@ Nyx::initData ()
                 GpuArray<amrex::Real,max_prob_param> prob_param;
                 prob_param_fill(prob_param);
                 prob_param_special_fill(prob_param);
-		comoving_type=std::round(prob_param[comoving_type_comp]);
+		comoving_type=int(std::round(prob_param[comoving_type_comp]));
 
                 prob_initdata_on_box(bx, fab_S_new, fab_D_new, geomdata, prob_param);
             }
@@ -296,7 +296,7 @@ Nyx::initData ()
                 GpuArray<amrex::Real,max_prob_param> prob_param;
                 prob_param_fill(prob_param);
                 prob_param_special_fill(prob_param);
-		comoving_type=std::round(prob_param[comoving_type_comp]);
+		comoving_type=int(std::round(prob_param[comoving_type_comp]));
 
                 prob_initdata_state_on_box(bx, fab_S_new, geomdata, prob_param);
             }
